Initial version of the software
This is the software for the RGB moodlight using PIC16F628 or PIC16F628A.
Adapt the included compiler headers to suit your build environment. If using XC8, then no change needs to be done.

If you are looking for compiled hex file ready to run with the given schematic, this is also included in the repository for your convenience.